+++ lua/obsidian/mappings.lua
@@ -14,6 +14,7 @@ M.smart_action = function()
   }
 end
 
+---@return obsidian.mappings.MappingConfig
 M.gf_passthrough = function()
   return {
     action = util.gf_passthrough,
@@ -21,11 +22,31 @@ M.gf_passthrough = function()
   }
 end
 
+---@return obsidian.mappings.MappingConfig
 M.toggle_checkbox = function()
   return {
     action = util.toggle_checkbox,
     opts = { buffer = true, desc = "Toggle Checkbox" },
   }
 end
+
+---@return obsidian.mappings.MappingConfig
+M.cycle_global = function()
+  return {
+    action = util.cycle_global,
+    opts = { buffer = true, desc = "Cycle file heading state" },
+  }
+end
+
+---@return obsidian.mappings.MappingConfig
+M.cycle = function()
+  return {
+    action = util.cycle,
+    opts = { buffer = true, desc = "Cycle heading state under the cursor" },
+  }
+end
+
+vim.keymap.set("n", "<Plug>(ObsidianCycle)", util.cycle)
+vim.keymap.set("n", "<Plug>(ObsidianCycleGlobal)", util.cycle_global)
 
 return M